fixes making this module actually run (bearly)
Rename "domain" to instance, because it's really more like that and the
fact that we use domain for an identifier is coincidence.

I forgot to add $mpm to the paramters of httpd::instance

We are now creating the directory where we want to put configuration
files (created in httpd::instance, used in also in httpd::vhost)

Templates now are trimmed, and wor: Thanks to Alex.

But only bearly: In the next step we'll use concat::fragment.
The reason for this is that multiple classes/types cannot manage
the same resource -- in this case, the same configuration file.

We will instead use a central concat::fragment type which all use to put
together a single file.